requirements.txt
Django==5.1.3
django-crispy-forms==2.1
django-filter==24.2
pillow==10.3.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt
dj-database-url
psycopg
gunicorn
psycopg2-binary